[Lib] Create Config.h
Create Config.h header for common configuration, and add the definition for F_CPU.

Include Config.h in TWI and UART drivers that relay on F_CPU definition. This Also fixes where defining F_CPU where the TWI and UART drivers are used in the application would have no effect on the drivers and the 1 MHz value provided as a fall back in the driver itself would always be used instead.